refactor: 구입 금액 받아오는 기능 분리
